single header and source file
